Crate yewtil

Utility library for the Yew frontend web framework.

All features:

"neq" - NeqAssign trait

"pure" - Pure components and function components.

"future" - Async support for Yew Messages

"fetch" - Wrapper that holds requests and responses.

"mrc_irc" - Ergonomic Rc pointers.

"lrc" - Linked-list Rc pointer.

"history" - History tracker

"store" - Global state with easy binding

Modules

future

ptr
Smart pointers for use within Yew.

store

Structs

History
Wrapper that keeps track of prior values that have been assigned to it.

Pure
Wrapper component for pure components.

Traits

NeqAssign
Blanket trait to provide a convenience method for assigning props in changed or updating values in update.

NeqAssignBy
Blanket trait to provide a convenience method for assigning props in changed or updating values in update.

PureComponent
Allows immutable components to be declared using a single struct and a single method.

Attribute Macros

function_component
Define a component with a function