Trait yewtil::future::LinkFuture

pub trait LinkFuture {
    type Message;
    fn callback_future<FN, FU, IN, M>(&self, function: FN) -> Callback<IN>
    where
        M: Into<Self::Message>,
        FU: Future<Output = M> + 'static,
        FN: Fn(IN) -> FU + 'static;

    fn send_future<F, M>(&self, future: F)
    where
        M: Into<Self::Message>,
        F: Future<Output = M> + 'static;

    fn send_future_batch<F>(&self, future: F)
    where
        F: Future<Output = Vec<Self::Message>> + 'static;
}

Trait that allows you to use ComponentLink and AgentLink to register futures.

Associated Types

type Message

Loading content...
Required methods

fn callback_future<FN, FU, IN, M>(&self, function: FN) -> Callback<IN> where
M: Into<Self::Message>,
FU: Future<Output = M> + 'static,
FN: Fn(IN) -> FU + 'static,

This method creates a Callback which returns a Future which returns a message to be sent back to the component's event loop.

Panics

If the future panics, then the promise will not resolve, and will leak.

fn send_future<F, M>(&self, future: F) where
M: Into<Self::Message>,
F: Future<Output = M> + 'static,

This method processes a Future that returns a message and sends it back to the component's loop.

Panics

If the future panics, then the promise will not resolve, and will leak.

fn send_future_batch<F>(&self, future: F) where
F: Future<Output = Vec<Self::Message>> + 'static,

Registers a future that resolves to multiple messages.

Panics

If the future panics, then the promise will not resolve, and will leak.

Loading content...