Struct yewtil::ptr::Mrc

pub struct Mrc<T> { /* fields omitted */ }

Mutable Reference Counted pointer

The Mrc has similar semantics to std::rc::Rc pointer, with notable differences that it does not support Weak pointers, it supports std::ops::DerefMut via the possibly allocating make_mut function, and that it can create immutable handles to its data (Irc).

This should make it just slightly more size efficient and performant than Rc, and should be more ergonomic to use than Rc given that you can mutably assign to it without much ceremony.

Passing Irc pointers to children guarantee that no intermediate component can modify the value behind the pointer. This makes it ideal for passing around configuration data where some components can ergonomicly "modify" and cheaply pass the pointers back to parent components, while other components can only read it.

Note

Assigning to an Mrc within Yew when you have passed shared copies of the ptr to child components, will always end up cloning the value stored in the Mrc. Rc makes this performance cost explicit, by making you use Rc::make_mut(). Because cloning is unavoidable in the context it was designed for, Mrc opts to provide nicer ergonomics around assignment.

Example

use yewtil::ptr::Mrc;

let mut mrc = Mrc::new(5);
*mrc = 10; // This just replaces the value because the mrc isn't shared.

assert_eq!(*mrc, 10);

let clone = mrc.clone();
*mrc = 20; // This operation clones the value and allocates space for it.

assert_eq!(*clone, 10);
assert_eq!(*mrc, 20);

Implementations

impl<T> Mrc<T>

pub fn new(value: T) -> Self

Allocates a value behind a Mrc pointer.

pub fn get_mut(&mut self) -> Option<&mut T>

Attempts to get a mutable reference to the wrapped value.

If the pointer is not shared, it will return Some, whereas if multiple Mrcs or Ircs point to the value, this will return None.

Example

use yewtil::ptr::Mrc;
let mut mrc = Mrc::new(0);
assert!(mrc.get_mut().is_some());

let _clone = mrc.clone();
assert!(mrc.get_mut().is_none());

pub fn try_unwrap(self) -> Result<T, Self>

Tries to extract the value from the Mrc, returning the Mrc if there is one or more other pointers to the value.

Example

use yewtil::ptr::Mrc;
let mrc = Mrc::new(0);

let clone = mrc.clone();
let mrc = mrc.try_unwrap().expect_err("Should not be able to unwrap");

std::mem::drop(clone);
let value = mrc.try_unwrap().expect("Should get value");

pub fn get_count(&self) -> usize

Gets the reference count of the Mrc.

An exclusive Mrc will have a count of 1. The count is incremented on any cloning action and is decremented when drop is called.

Example

use yewtil::ptr::Mrc;
let mrc = Mrc::new(0);
assert_eq!(mrc.get_count(), 1);

let _clone = mrc.clone();
assert_eq!(mrc.get_count(), 2);

std::mem::drop(_clone);
assert_eq!(mrc.get_count(), 1);

pub fn is_exclusive(&self) -> bool

Returns true if no other pointers to the value exist.

use yewtil::ptr::Mrc;
let mrc = Mrc::new(0);
assert!(mrc.is_exclusive());

let _clone = mrc.clone();
assert!(!mrc.is_exclusive());

std::mem::drop(_clone);
assert!(mrc.is_exclusive());

pub fn irc(&self) -> Irc<T>

Returns an immutable reference counted pointer, pointing to the same value and reference count.

Example

use yewtil::ptr::{Mrc, Irc};
let mrc: Mrc<usize> =  Mrc::new(0);
let _irc: Irc<usize> = mrc.irc();

assert!(!mrc.is_exclusive());

pub fn into_irc(self) -> Irc<T>

Converts this Mrc into an Irc.

Example

use yewtil::ptr::{Mrc, Irc};
let mrc: Mrc<usize> =  Mrc::new(0);
let irc: Irc<usize> = mrc.into_irc();

assert!(irc.is_exclusive());

pub fn ptr_eq(lhs: &Self, rhs: &Self) -> bool

Checks pointers for equality.

Example

use yewtil::ptr::Mrc;
let mrc1 = Mrc::new(0);
let mrc2 = Mrc::new(0);
assert_eq!(mrc1, mrc2);
assert!(!Mrc::ptr_eq(&mrc1, &mrc2))

impl<T: Clone> Mrc<T>

pub fn make_mut(&mut self) -> &mut T

Returns a mutable reference to the value if it has exclusive access. If it does not have exclusive access, it will make a clone of the data to acquire exclusive access.

Example

let mut mrc: Mrc<usize> = Mrc::new(0);

let _mut_ref: &mut usize = mrc.make_mut();
assert_eq!(mrc.get_count(), 1);

let clone = mrc.clone();
assert_eq!(mrc.get_count(), 2);

let _mut_ref: &mut usize = mrc.make_mut();
assert_eq!(mrc.get_count(), 1);
assert!(!Mrc::ptr_eq(&mrc, &clone))

pub fn unwrap_clone(self) -> T

Consumes the Mrc and returns the value from the Mrc if it is not shared or clones the value if another Mrc or Irc has access to it.

pub fn clone_inner(&self) -> T

Clones the value wrapped by the Mrc..