Module yewtil::store

Structs

ReadOnly
A wrapper ensuring state observers can only borrow the state immutably

StoreWrapper
Hides the full context Agent from a Store and does the boring data wrangling logic

Traits

Bridgeable
Allows us to communicate with a store

Store
A functional state wrapper, enforcing a unidirectional data flow and consistent state to the observers.